rm exsiccati number from dwc:recordNumber field
From Ed Gilbert:  Exsiccati number is NOT the collector’s number
(recordNumber). Typically, collectors don’t assign a collector’s number
when the specimen is an exsiccati, at least for lichens. And,
Exsiicati titles tend to be latinized (see
http://lichenportal.org/portal/collections/exsiccati/index.php).
Typically, exsiccati numbers tend to be in front of the scientific name,
but not always.